Fix sample name
